--- .gitignore
@@ -1,4 +1,0 @@
-.env
-07hack
-.hoge
-hackthon07
