--- frontend/src/page/signup.tsx
@@ -1,10 +1,0 @@
-// Homeコンポーネントの作成
-import React from "react";
-import Signup from "../feature/auth/Signup";
-
-// Homeコンポーネントの作成
-const SignupPage = () => {
-  return <Signup />;
-};
-
-export default SignupPage;
